Add some basic tracing to ociplex
After testing various logging and tracing options, I settled for
tokio::tracing, mostly because it's also used in youki. However, it
seems that there is no standard crate yet to connect it to the syslog.

The implementation here is only partial, it does not yet honor the
`--log` option to select a log file, and the option to log to the
syslog.

Signed-off-by: Christophe de Dinechin <dinechin@redhat.com>